module_sf_ruclsm imprecision workaround
The module_sf_ruclsm has comparisons to zero, which break the code
when numbers are very close to 0, such as 1e-322. This happens with
the gfortran compiler when compled -DDEBUG=ON since the option to
truncate subnormal numbers is turned off.

The workaround is to truncate subnormal numbers manually, for the
two variables that cause trouble.