BUG: Match ozone diagnostics to terms of the parametrization
I want the ozone tendency diagnostics to have the same units as each other (it looks like kg/kg here, presumably there's a step in the driver to accumulate these and divide by the interval of accumulation).  

I would also like each diagnostic to correspond to a single term in the parametrization, rather than one-and-a-half and then a different half.