Make ozone photochem diagnostics match parametrization terms
The ozone photochemistry parametrization is a first-order Taylor expansion of the net ozone production in three variables.  This commit changes the four diagnostics the parametrization manages to match the four terms in the Taylor expansion.

When comparing to the implementation in lines 301-304, note that it is implicit in the model ozone mixing ratio, so line 301 has the first term and the second half of the second term, while the first half of the second term is part of the denominator on line 304.

I could use ozi or ozib for an explicit approximation to that term, but we already calculated the value needed for the implicit calculation actually used for the update.